Changed edge weight type from int to float

+++ Sem2/AlgorytmyIStrukturyDanych/Lista 9/MyGraph/src/GraphPackage/MyGraphInterface.java
@@ -3,5 +3,5 @@ package GraphPackage;
 public interface MyGraphInterface <V>
 {
     void addVertex(V vertex);
-    public void addEdge(V source, V destination,int weight);
+    public void addEdge(V source, V destination, float weight);
 }
